third_party/WebKit/LayoutTests/fast/css/invalidation/independent-inheritance-fast-path-inherit-keyword.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="outer" style="visibility: hidden;">
    <div id="inner" style="visibility: inherit;"></div>
</div>
<script>
test(function(t)
{
    assert_equals(getComputedStyle(inner).visibility, "hidden");

    outer.offsetTop; // Force recalc.
    outer.style.visibility = "visible";
    assert_equals(internals.updateStyleAndReturnAffectedElementCount(), 1);
    assert_equals(getComputedStyle(inner).visibility, "visible");

}, "Changing visibility, an independent inherited property, propagates correctly when using the 'inherit' keyword.");
</script>
